added a lost function
